Refactored ketchup-fld to kup-fld
