+++ packages/ketchup/src/components.d.ts
@@ -25,7 +25,7 @@ import {
 import {
   KetchupFldChangeEvent,
   KetchupFldSubmitEvent,
-} from './components/ketchup-fld/ketchup-fld-declarations';
+} from './components/ketchup-fld/kup-fld-declarations';
 import {
   ElementOffset,
 } from './utils/offset';
@@ -124,7 +124,7 @@ export namespace Components {
     'totals'?: TotalsMap;
   }
 
-  interface KetchupFld {
+  interface KupFld {
     /**
     * Data the FLD must parse to fully be configured. It must be either an Object or a JSON parsable string
     */
@@ -138,7 +138,7 @@ export namespace Components {
     */
     'getCurrentValue': () => Promise<string | object>;
   }
-  interface KetchupFldAttributes extends StencilHTMLAttributes {
+  interface KupFldAttributes extends StencilHTMLAttributes {
     /**
     * Data the FLD must parse to fully be configured. It must be either an Object or a JSON parsable string
     */
@@ -620,8 +620,7 @@ export namespace Components {
 
 declare global {
   interface StencilElementInterfaces {
-    'KupDataTable': Components.KupDataTable;
-    'KetchupFld': Components.KetchupFld;
+    'KupFld': Components.KupFld;
     'KetchupHtml': Components.KetchupHtml;
     'KupPaginator': Components.KupPaginator;
     'KetchupPortalInstance': Components.KetchupPortalInstance;
@@ -637,8 +636,7 @@ declare global {
   }
 
   interface StencilIntrinsicElements {
-    'kup-data-table': Components.KupDataTableAttributes;
-    'ketchup-fld': Components.KetchupFldAttributes;
+    'kup-fld': Components.KupFldAttributes;
     'ketchup-html': Components.KetchupHtmlAttributes;
     'kup-paginator': Components.KupPaginatorAttributes;
     'ketchup-portal-instance': Components.KetchupPortalInstanceAttributes;
@@ -660,10 +658,10 @@ declare global {
     new (): HTMLKupDataTableElement;
   };
 
-  interface HTMLKetchupFldElement extends Components.KetchupFld, HTMLStencilElement {}
-  var HTMLKetchupFldElement: {
-    prototype: HTMLKetchupFldElement;
-    new (): HTMLKetchupFldElement;
+  interface HTMLKupFldElement extends Components.KupFld, HTMLStencilElement {}
+  var HTMLKupFldElement: {
+    prototype: HTMLKupFldElement;
+    new (): HTMLKupFldElement;
   };
 
   interface HTMLKetchupHtmlElement extends Components.KetchupHtml, HTMLStencilElement {}
@@ -739,8 +737,7 @@ declare global {
   };
 
   interface HTMLElementTagNameMap {
-    'kup-data-table': HTMLKupDataTableElement
-    'ketchup-fld': HTMLKetchupFldElement
+    'kup-fld': HTMLKupFldElement
     'ketchup-html': HTMLKetchupHtmlElement
     'kup-paginator': HTMLKupPaginatorElement
     'ketchup-portal-instance': HTMLKetchupPortalInstanceElement
@@ -756,8 +753,7 @@ declare global {
   }
 
   interface ElementTagNameMap {
-    'kup-data-table': HTMLKupDataTableElement;
-    'ketchup-fld': HTMLKetchupFldElement;
+    'kup-fld': HTMLKupFldElement;
     'ketchup-html': HTMLKetchupHtmlElement;
     'kup-paginator': HTMLKupPaginatorElement;
     'ketchup-portal-instance': HTMLKetchupPortalInstanceElement;
