+++ packages/ketchup/src/components/ketchup-fld/kup-fld.scss
@@ -10,7 +10,7 @@
 }
 
 // Base class and its derived
-.ketchup-fld {
+.kup-fld {
   $rightClass: #{&}--right;
 
   // the top div container
