+++ packages/ketchup/src/components/ketchup-fld/readme.md
@@ -1,4 +1,4 @@
-# ketchup-fld
+# kup-fld
