+++ packages/ketchup/src/index.html
@@ -77,12 +77,12 @@
         ></ketchup-text-input>
         <kup-combo is-clearable use-portal="true"></kup-combo>
         <h4>FLD 1</h4>
-        <ketchup-fld id="kfld-1"></ketchup-fld>
+        <kup-fld id="kfld-1"></kup-fld>
         <h4>FLD 2</h4>
-        <ketchup-fld id="kfld-2"></ketchup-fld>
-        <ketchup-fld id="kfld-3"></ketchup-fld>
-        <ketchup-fld id="kfld-4"></ketchup-fld>
-        <ketchup-fld id="fld-radio-initialValue"></ketchup-fld>
+        <kup-fld id="kfld-2"></kup-fld>
+        <kup-fld id="kfld-3"></kup-fld>
+        <kup-fld id="kfld-4"></kup-fld>
+        <kup-fld id="fld-radio-initialValue"></kup-fld>
 
         <script>
             let once = true;
